implemented image segmentation image showcase, to be edited
